requirements.txt
appnope==0.1.0
asgiref==3.3.4
backcall==0.1.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
decorator==4.3.2
dj-database-url==0.5.0
django-cors-headers==3.7.0
django-extensions==2.1.5
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
Flask==1.1.2
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
MarkupSafe==1.1.1
parso==0.8.1
peewee==3.14.2
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.8
ptyprocess==0.6.0
pycodestyle==2.4.0
Pygments==2.3.1
PyJWT==1.7.1
pytz==2018.7
PyYAML==5.4.1
redgreenunittest==0.1.1
requests==2.25.1
six==1.12.0
sqlparse==0.4.1
traitlets==4.3.2
urllib3==1.26.4
wcwidth==0.1.7
Werkzeug==1.0.1
whitenoise==5.2.0
